Tested runtime speed to see if things would work out okay with js -> wasm -> js and it turned out just fine
Wow did not expect this but after the first call to wasm it ended up
being ~100x faster 0.02ms per call than join-monster which was 1.7ms
(with a similar simple query) when the data is prepared before hand. Do
note that this is not the speed of the SQL query running, but the speed
at which the SQL query string is created which are two totally different
things. The first call to this library has to do some initialization and
is actually more like 1.5x slower (2.5ms) than join monster (1.7ms), but
each subsequent call is up to 100x faster! I figured passing data
through wasm would be incredibly slow and possibly too slow for this
library to function well, but I was wrong woot!